requirements.txt
openai
python-dotenv
typer[all]
rich
pytest
google-generativeai
pytest-asyncio 